Platform4.2.0 (#420)
* Update to new Arduino framework

* Added lib dependency on arduinoWebSockets git repo

* Tab in platformio.ini

* Working with new ESP32 platform 4.2.0

* Fixed rebase botch

* Improved file handling for LittleFS

1. The "filename too long" message has been changed to a log_debug()
   that shows the filename.  That works for all filesystems and can
   be turned on and off with $Message/Level
2. There was a bug in $localfs/delete that failed to delete LittleFS
   files because the file was open.  Fixed by closing the file first.

* Added file LocalFS.h

+++ FluidNC/src/Stepper.cpp
@@ -79,9 +79,9 @@ typedef struct {
 static stepper_t st;
 
 // Step segment ring buffer indices
-static volatile uint8_t segment_buffer_tail;
-static volatile uint8_t segment_buffer_head;
-static uint8_t          segment_next_head;
+static volatile uint32_t segment_buffer_tail;
+static volatile uint32_t segment_buffer_head;
+static uint32_t          segment_next_head;
 
 // Pointers for the step segment being prepped from the planner buffer. Accessed only by the
 // main program. Pointers may be planning segments or planner blocks ahead of what being executed.
@@ -269,7 +269,6 @@ void IRAM_ATTR Stepper::pulse_func() {
 
 // enabled. Startup init and limits call this function but shouldn't start the cycle.
 void Stepper::wake_up() {
-    //log_info("st_wake_up");
     // Cancel any pending stepper disable
     protocol_cancel_disable_steppers();
     // Enable stepper drivers.
